Added .travis.yml file and badge image.

+++ test/clean-test-environment.sh
@@ -1,3 +1,6 @@
 #!/bin/bash
 
-rm -rf test-website
+if [ -d "test-website" ]
+then
+  rm -rf test-website
+fi
